requirements.txt
asgiref==3.9.1
Django==4.2.24
django-cors-headers==4.8.0
djangorestframework==3.16.1
mysqlclient==2.2.7
PyMySQL==1.1.2
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
